items_price module add and code refactor
